split ItemsWeaponsSpriteFactory into ItemsSpriteFactory, WeaponsSpriteFactory, ProjectileSpriteFactory, and (temporarily) MiscSpriteFactory. Also updated all instances where ItemWeaponsSpritrFactory was used.
