+++ sprint0/sprint0/Items/BlueRupee.cs
@@ -1,0 +1,38 @@
+﻿using System;
+using System.Collections.Generic;
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
+
+//Author: Angela Li
+namespace sprint0
+{
+    public class BlueRupee : IItem
+    {
+        public int PickedUpDuration { get; set; }
+        private readonly int maxPickedUpDuration = 40;
+        public Rectangle Location { get; set; }
+        public Texture2D Texture { get; set; }
+        private Rectangle source;
+        private readonly int xOffset = 72, yOffset = 16, width = 8, height = 16;
+
+        public BlueRupee(Texture2D texture, Vector2 location)
+        {
+            Location = new Rectangle((int)location.X, (int)location.Y, (int)(width * Game1.Scale), (int)(height * Game1.Scale));
+            Texture = texture;
+
+            //load sprites
+            source = new Rectangle(xOffset, yOffset, width, height);
+        }
+
+        public void Draw(SpriteBatch spriteBatch)
+        {
+            if (PickedUpDuration < maxPickedUpDuration)
+                spriteBatch.Draw(Texture, Location, source, Color.White);
+        }
+
+        public void Update()
+        {
+            if (PickedUpDuration >= 0) PickedUpDuration++;
+        }
+    }
+}
